Added 2 booleans to the mail model for deletefrom; inbox / outbox, added a check to the delete function in the mail controller so that only if both bools are true is it deleted. I will send you a change script. Added a Context to all notifications so there should no longer be empty "[]"

+++ tags/Fall 2011/web/OSBLE/Controllers/MailController.cs
@@ -356,6 +356,8 @@ namespace OSBLE.Controllers
                         newMail.Message = mail.Message;
                         newMail.ThreadID = threadID;
                         newMail.ContextID = mail.ContextID;
+                        newMail.DeleteFromInbox = false;
+                        newMail.DeleteFromOutbox = false;
 
                         //need to create the mail before we can send the notification and set the threadID
                         db.Mails.Add(newMail);
@@ -391,8 +393,29 @@ namespace OSBLE.Controllers
 
             if (mail.ToUserProfile == currentUser)
             {
-                db.Mails.Remove(mail);
-                db.SaveChanges();
+                if (mail.DeleteFromOutbox == true)
+                {
+                    db.Mails.Remove(mail);
+                    db.SaveChanges();
+                }
+                else
+                {
+                    mail.DeleteFromInbox = true;
+                    db.SaveChanges();
+                }
+            }
+            else if (mail.FromUserProfile == currentUser)
+            {
+                if (mail.DeleteFromInbox == true)
+                {
+                    db.Mails.Remove(mail);
+                    db.SaveChanges();
+                }
+                else
+                {
+                    mail.DeleteFromOutbox = true;
+                    db.SaveChanges();
+                }
             }
             else
             {
